debug hover enhancements
