--- src/commands/execute-cql.ts
@@ -1,728 +1,0 @@
-import * as fse from 'fs-extra';
-import { glob } from 'glob';
-import { ParseError, parse as parseJsonc } from 'jsonc-parser';
-import * as fs from 'node:fs';
-import {
-  commands,
-  debug,
-  ExtensionContext,
-  Position,
-  ProgressLocation,
-  Range,
-  TextEditor,
-  Uri,
-  window,
-  workspace,
-} from 'vscode';
-import { Utils } from 'vscode-uri';
-import { Commands } from '../commands/commands';
-import { ExecuteCqlResponse, ExpressionResult, executeCql } from '../cql-service/cqlService.executeCql';
-import { CqlLibrary, CqlProject, CqlProjectEvents } from '../model/cqlProject';
-import { CqlSolution } from '../model/cqlSolution';
-import { extractLibraryVersion, toGlobPath } from '../helpers/fileHelper';
-import { resolveParameters } from '../helpers/parametersHelper';
-import * as log from '../log-services/logger';
-import { CqlParametersConfig, ParameterEntry, ResultParameterEntry } from '../model/parameters';
-import { getMeasureReportData, getTestCases, TestCase, TestCaseExclusion } from '../model/testCase';
-
-let _context: ExtensionContext | undefined;
-
-export interface TestCaseResult {
-  executedAt: string;
-  libraryName: string;
-  testCaseName: string | null;
-  testCaseDescription: string | null;
-  /** All parameters for this evaluation — config-supplied and CQL-declared defaults — ordered config first, defaults appended. Differentiate by the `source` field. */
-  parameters: ResultParameterEntry[];
-  results: ExpressionResult[];
-  errors: string[];
-}
-
-interface CqlPaths {
-  libraryDirectoryPath: Uri;
-  projectDirectoryPath: Uri;
-  optionsPath: Uri;
-  resultDirectoryPath: Uri;
-  terminologyDirectoryPath: Uri;
-  testConfigPath: Uri;
-  testDirectoryPath: Uri;
-}
-
-export interface TestConfig {
-  testCasesToExclude: TestCaseExclusion[];
-  parameters?: CqlParametersConfig;
-  resultFormat?: 'individual' | 'flat';
-}
-
-const DEFAULT_TEST_CONFIG: TestConfig = { testCasesToExclude: [], resultFormat: 'flat' }
-
-export function register(context: ExtensionContext): void {
-  _context = context;
-
-  // Track when library shells are loaded so the select-libraries command can be enabled in menus.
-  // Uses libraryShellsLoaded to handle the race where LOADED fires before register() runs.
-  const solution = CqlSolution.getCurrent();
-  commands.executeCommand('setContext', 'cql.solutionLoaded', false);
-  const total = solution.projects.length;
-  if (total === 0) {
-    commands.executeCommand('setContext', 'cql.solutionLoaded', true);
-  } else {
-    let loadedCount = solution.projects.filter(p => p.libraryShellsLoaded).length;
-    if (loadedCount >= total) {
-      commands.executeCommand('setContext', 'cql.solutionLoaded', true);
-    } else {
-      for (const project of solution.projects) {
-        if (!project.libraryShellsLoaded) {
-          project.once(CqlProjectEvents.LOADED, () => {
-            if (++loadedCount >= total) {
-              commands.executeCommand('setContext', 'cql.solutionLoaded', true);
-            }
-          });
-        }
-      }
-    }
-  }
-
-  context.subscriptions.push(
-    commands.registerCommand(Commands.EXECUTE_CQL_COMMAND, async (uri: Uri) => {
-      executeCQLFile(uri);
-    }),
-    commands.registerCommand(Commands.EXECUTE_CQL_COMMAND_SELECT_LIBRARIES, async (uri: Uri) => {
-      selectLibraries();
-    }),
-    commands.registerCommand(Commands.EXECUTE_CQL_COMMAND_SELECT_TEST_CASES, async (uri: Uri) => {
-      selectTestCases(uri);
-    }),
-    commands.registerCommand(Commands.DEBUG_TEST_CASE_COMMAND, async (uri: Uri) => {
-      debugTestCase(uri);
-    }),
-  );
-}
-
-export async function selectLibraries(): Promise<void> {
-  // Use the active editor's URI to scope to the right project; fall back to
-  // the first project if no editor is open or the file is not in any project.
-  const activeUri = window.activeTextEditor?.document.uri;
-  const solution = CqlSolution.getCurrent();
-  const fallbackProject = solution.projects[0];
-  const anchorUri = activeUri
-    ? (solution.findProjectForUri(activeUri) ? activeUri : fallbackProject && Uri.file(fallbackProject.igRoot))
-    : (fallbackProject && Uri.file(fallbackProject.igRoot));
-  if (!anchorUri) {
-    window.showErrorMessage('No CQL project found in this workspace.');
-    return;
-  }
-  const cqlPaths = getCqlPaths(anchorUri);
-  if (!cqlPaths) {
-    window.showErrorMessage('Unable to determine needed CQL Paths.');
-    return;
-  }
-
-  const libraries = getLibraries(cqlPaths.libraryDirectoryPath);
-  const quickPickItems = libraries
-    .map(uri => ({ label: Utils.basename(uri), uri }))
-    .sort((a, b) => a.label.localeCompare(b.label));
-
-  const quickPick = window.createQuickPick<{ label: string; uri: Uri }>();
-  if (quickPickItems.length > 0) {
-    quickPick.items = quickPickItems;
-    quickPick.canSelectMany = true;
-
-    const stateKey = 'selectLibraries.selections';
-    const savedSelections = _context?.workspaceState.get<string[]>(stateKey) ?? [];
-    if (savedSelections.length > 0) {
-      quickPick.selectedItems = quickPick.items.filter(item =>
-        savedSelections.includes(item.label),
-      );
-    }
-
-    quickPick.show();
-
-    quickPick.onDidAccept(async () => {
-      const selected = [...quickPick.selectedItems];
-      _context?.workspaceState.update(stateKey, selected.map(item => item.label));
-      quickPick.hide();
-
-      await window.withProgress(
-        {
-          location: ProgressLocation.Notification,
-          title: 'Executing CQL Libraries',
-          cancellable: true,
-        },
-        async (progress, token) => {
-          const total = selected.length;
-          const batchStart = Date.now();
-          let completed = 0;
-          for (let i = 0; i < total; i++) {
-            if (token.isCancellationRequested) {
-              break;
-            }
-            const item = selected[i];
-            progress.report({
-              message: `(${i + 1}/${total}) ${item.label}`,
-              increment: (1 / total) * 100,
-            });
-            const libStart = Date.now();
-            try {
-              await executeCQLFile(item.uri, undefined, false, undefined, false);
-              log.info(`[PERF] ${item.label}: ${((Date.now() - libStart) / 1000).toFixed(1)}s`);
-              completed++;
-            } catch (e) {
-              log.error(`Error executing CQL for ${item.label}`, e);
-              window.showErrorMessage(
-                `Failed to execute ${item.label}: ${e instanceof Error ? e.message : String(e)}`,
-              );
-            }
-            await new Promise(resolve => setTimeout(resolve, 500));
-          }
-          const batchElapsed = ((Date.now() - batchStart) / 1000).toFixed(1);
-          log.info(`[PERF] selectLibraries total (${total} libraries): ${batchElapsed}s`);
-          const msg =
-            completed === total
-              ? `CQL execution complete — ${total} ${total === 1 ? 'library' : 'libraries'} (${batchElapsed}s)`
-              : `CQL execution cancelled — ${completed}/${total} libraries (${batchElapsed}s)`;
-          window.showInformationMessage(msg);
-        },
-      );
-    });
-  }
-}
-
-export async function selectTestCases(cqlFileUri: Uri): Promise<void> {
-  const cqlPaths = getCqlPaths(cqlFileUri);
-  if (!cqlPaths) {
-    const msg = 'Unable to resolve needed CQL project paths.';
-    log.error(msg);
-    window.showErrorMessage(msg);
-    return;
-  }
-
-  const quickPick = window.createQuickPick();
-  const libraryName = Utils.basename(cqlFileUri).replace('.cql', '').split('-')[0];
-  const testConfig = loadTestConfig(cqlPaths.testConfigPath);
-  const excludedTestCases = getExcludedTestCases(libraryName, testConfig.testCasesToExclude);
-  const testCases = getTestCases(
-    cqlPaths.testDirectoryPath,
-    libraryName,
-    Array.from(excludedTestCases.keys()),
-  );
-  const namedTestCases = testCases.filter(
-    (testCase): testCase is Required<TestCase> => testCase.name !== undefined,
-  );
-  const quickPickItems = namedTestCases.map(testCase => ({
-    label: testCase.name,
-    detail: getMeasureReportData(testCase.path)?.description,
-  }));
-
-  if (quickPickItems.length > 0) {
-    quickPick.items = quickPickItems;
-    quickPick.canSelectMany = true;
-    quickPick.matchOnDetail = true;
-
-    const stateKey = `selectTestCases.selections.${libraryName}`;
-    const savedSelections = _context?.workspaceState.get<string[]>(stateKey) ?? [];
-    if (savedSelections.length > 0) {
-      quickPick.selectedItems = quickPick.items.filter(item =>
-        savedSelections.includes(item.label),
-      );
-    }
-
-    quickPick.show();
-
-    quickPick.onDidAccept(() => {
-      const selectedLabels = quickPick.selectedItems.map(item => item.label);
-      _context?.workspaceState.update(stateKey, selectedLabels);
-
-      const selected = namedTestCases.filter(testCase =>
-        selectedLabels.some(label => label === testCase.name),
-      );
-      quickPick.hide();
-      executeCQLFile(cqlFileUri, selected);
-    });
-  }
-}
-
-export async function debugTestCase(cqlFileUri: Uri): Promise<void> {
-  const cqlPaths = getCqlPaths(cqlFileUri);
-  if (!cqlPaths) {
-    window.showErrorMessage('Unable to resolve needed CQL project paths.');
-    return;
-  }
-
-  const libraryName = Utils.basename(cqlFileUri).replace('.cql', '').split('-')[0];
-
-  const allTestCases = getTestCases(cqlPaths.testDirectoryPath, libraryName, []);
-  const namedTestCases = allTestCases.filter(
-    (tc): tc is Required<TestCase> => tc.name !== undefined,
-  );
-
-  if (namedTestCases.length === 0) {
-    window.showInformationMessage('No test cases found.');
-    return;
-  }
-
-  const quickPickItems = namedTestCases.map(tc => ({
-    label: tc.name,
-    testCase: tc,
-  }));
-
-  const quickPick = window.createQuickPick<{ label: string; testCase: TestCase }>();
-  quickPick.items = quickPickItems;
-  quickPick.canSelectMany = false;
-  quickPick.placeholder = 'Select a test case to debug';
-  quickPick.title = `Debug — ${libraryName}`;
-
-  const stateKey = `debugTestCase.selections.${libraryName}`;
-  const saved = _context?.workspaceState.get<string>(stateKey);
-  if (saved) {
-    const match = quickPickItems.find(item => item.label === saved);
-    if (match) {
-      quickPick.selectedItems = [match];
-      quickPick.activeItems = [match];
-    }
-  }
-
-  quickPick.show();
-
-  quickPick.onDidAccept(async () => {
-    const selected = quickPick.selectedItems[0];
-    quickPick.hide();
-    if (!selected) return;
-
-    await _context?.workspaceState.update(stateKey, selected.label);
-
-    const cqlSource = fs.readFileSync(cqlFileUri.fsPath, 'utf-8');
-    const fhirVersion = getFhirVersion(cqlSource) ?? 'R4';
-
-    await debug.startDebugging(
-      workspace.getWorkspaceFolder(cqlFileUri),
-      {
-        type: 'cql',
-        request: 'launch',
-        name: `Debug ${libraryName} — ${selected.label}`,
-        libraryUri: cqlFileUri.toString(),
-        libraryName,
-        fhirVersion,
-        testCaseName: selected.label,
-        testCaseUri: selected.testCase.path!.toString(),
-        terminologyUri: cqlPaths.terminologyDirectoryPath?.toString(),
-        rootDir: cqlPaths.projectDirectoryPath?.toString(),
-        optionsPath: cqlPaths.optionsPath?.toString(),
-      },
-    );
-  });
-}
-
-export async function executeCQLFile(
-  cqlFileUri: Uri,
-  testCases: Array<TestCase> | undefined = undefined,
-  showProgress: boolean = true,
-  resultFormatOverride?: string,
-  showCompletion: boolean = true,
-): Promise<void> {
-  if (!fs.existsSync(cqlFileUri.fsPath)) {
-    window.showInformationMessage('No library content found. Please save before executing.');
-    return;
-  }
-
-  const cqlPaths = getCqlPaths(cqlFileUri);
-  if (!cqlPaths) {
-    window.showErrorMessage('Unable to determine needed CQL Paths.');
-    return;
-  }
-
-  const libraryName = Utils.basename(cqlFileUri).replace('.cql', '').split('-')[0];
-  const libraryDisplayName = Utils.basename(cqlFileUri).replace('.cql', '');
-
-  const testConfig = loadTestConfig(cqlPaths.testConfigPath);
-  const excludedTestCases = getExcludedTestCases(libraryName, testConfig.testCasesToExclude);
-
-  // When no test cases are provided we are about to scan disk for them.  Wait for the
-  // in-memory model to finish loading first so we never execute against an empty set
-  // during initial workspace startup.  No-op if the library is already loaded.
-  if (testCases === undefined) {
-    const project = CqlSolution.getCurrent().findProjectForUri(cqlFileUri);
-    const library = project?.Libraries.find(lib => lib.uri.fsPath === cqlFileUri.fsPath);
-    if (library && project) {
-      await waitForTestCasesLoaded(library, project, showProgress);
-    }
-  }
-
-  const effectiveTestCases: Array<TestCase> =
-    testCases ??
-    getTestCases(cqlPaths.testDirectoryPath, libraryName, Array.from(excludedTestCases.keys()));
-
-  // We didn't find any test cases, so we'll just execute an empty one
-  if (effectiveTestCases.length === 0) {
-    effectiveTestCases.push({});
-  }
-
-  const cqlSource = fs.readFileSync(cqlFileUri.fsPath, 'utf-8');
-  const determinedFhirVersion = getFhirVersion(cqlSource);
-  const fhirVersion: string = determinedFhirVersion ?? 'R4';
-  if (!determinedFhirVersion) {
-    window.showInformationMessage('Unable to determine version of FHIR used. Defaulting to R4.');
-  }
-  const libraryVersion = extractLibraryVersion(cqlSource);
-
-  const resultFormat = resultFormatOverride
-    ?? testConfig.resultFormat
-
-  const doExecute = () =>
-    executeCql(
-      cqlFileUri,
-      effectiveTestCases,
-      cqlPaths.terminologyDirectoryPath,
-      fhirVersion,
-      cqlPaths.optionsPath,
-      cqlPaths.projectDirectoryPath,
-      undefined,
-      testConfig.parameters,
-      libraryVersion,
-    );
-
-  const startExecution = Date.now();
-
-
-  let response: ExecuteCqlResponse | undefined;
-  if (showProgress) {
-    response = await window.withProgress(
-      {
-        location: ProgressLocation.Notification,
-        title: `Executing CQL: ${libraryDisplayName}`,
-        cancellable: false,
-      },
-      async progress => {
-        progress.report({ message: 'Running...' });
-        return doExecute();
-      },
-    );
-  } else {
-    response = await doExecute();
-  }
-
-  const endExecution = Date.now();
-  const elapsedSeconds = (endExecution - startExecution) / 1000;
-
-  if (resultFormat === 'individual') {
-    if (response) {
-      writeIndividualResultFiles(
-        libraryName,
-        libraryVersion,
-        effectiveTestCases,
-        response,
-        cqlPaths.resultDirectoryPath,
-        startExecution,
-        testConfig.parameters,
-      );
-    }
-    if (showCompletion && response) {
-      if (effectiveTestCases.length === 1) {
-        const patientId = effectiveTestCases[0].name ?? 'no-context';
-        const outputPath = Utils.resolvePath(
-          cqlPaths.resultDirectoryPath,
-          libraryName,
-          `TestCaseResult-${patientId}.json`,
-        );
-        const textDocument = await workspace.openTextDocument(outputPath);
-        await window.showTextDocument(textDocument);
-      } else {
-        const count = effectiveTestCases.length;
-        window.showInformationMessage(
-          `CQL execution complete — ${count} test cases written (${elapsedSeconds.toFixed(1)}s)`,
-        );
-      }
-    }
-  } else {
-    const outputPath = Utils.resolvePath(cqlPaths.resultDirectoryPath, `${libraryName}.txt`);
-    // Ensure the result directory and file exist for first-run cases.
-    // Do NOT truncate here if the file is already open in VS Code — that would bump the on-disk
-    // mtime and trigger a "file is newer" conflict when we later call textDocument.save().
-    // The textEditor.edit() delete below clears stale in-memory content on every run.
-    fse.ensureDirSync(cqlPaths.resultDirectoryPath.fsPath);
-    if (!fs.existsSync(outputPath.fsPath)) {
-      fs.writeFileSync(outputPath.fsPath, '');
-    }
-    const textDocument = await workspace.openTextDocument(outputPath);
-    const textEditor = await window.showTextDocument(textDocument);
-    // Clear any existing in-memory content before inserting new results.
-    // This handles repeated runs where VS Code still holds the previous run's content.
-    await textEditor.edit(
-      editBuilder => {
-        editBuilder.delete(
-          new Range(
-            new Position(0, 0),
-            textDocument.lineAt(Math.max(0, textDocument.lineCount - 1)).range.end,
-          ),
-        );
-      },
-      { undoStopBefore: false, undoStopAfter: false },
-    );
-
-    const cqlMessage = `CQL: ${cqlPaths.libraryDirectoryPath.fsPath}`;
-    const terminologyMessage = fs.existsSync(cqlPaths.terminologyDirectoryPath.fsPath)
-      ? `Terminology: ${cqlPaths.terminologyDirectoryPath.fsPath}`
-      : `No terminology found at ${cqlPaths.terminologyDirectoryPath.fsPath}. Evaluation may fail if terminology is required.`;
-
-    const testMessage: string[] = [];
-    if (effectiveTestCases.length === 1 && !effectiveTestCases[0].name) {
-      testMessage.push(
-        `No data found at ${cqlPaths.testDirectoryPath.fsPath}. Evaluation may fail if data is required.`,
-      );
-    } else {
-      testMessage.push(`Test cases:`);
-      for (const p of effectiveTestCases) {
-        testMessage.push(`${p.name} - ${p.path?.fsPath}`);
-      }
-    }
-
-    if (excludedTestCases.size > 0) {
-      testMessage.push('\nExcluded test cases:');
-      for (const [testCase, reason] of excludedTestCases.entries()) {
-        testMessage.push(`${testCase} - ${reason}`);
-      }
-    }
-
-    await insertLineAtEnd(textEditor, cqlMessage);
-    await insertLineAtEnd(textEditor, terminologyMessage);
-    await insertLineAtEnd(textEditor, `${testMessage.join('\n')}\n`);
-
-    if (response) {
-      await insertLineAtEnd(textEditor, formatResponse(response));
-    }
-    await insertLineAtEnd(
-      textEditor,
-      `\nelapsed: ${elapsedSeconds.toString()} seconds\n`,
-    );
-    await textDocument.save();
-  }
-}
-
-export function writeIndividualResultFiles(
-  libraryName: string,
-  libraryVersion: string | undefined,
-  testCases: Array<TestCase>,
-  response: ExecuteCqlResponse,
-  resultDirectoryPath: Uri,
-  executedAt: number,
-  parametersConfig?: CqlParametersConfig,
-): void {
-  const executedAtStr = new Date(executedAt).toISOString();
-
-  for (let i = 0; i < response.results.length; i++) {
-    const libraryResult = response.results[i];
-    const testCase = testCases[i];
-    const testCaseName = testCase?.name ?? null;
-    const testCaseDescription = testCase?.path
-      ? (getMeasureReportData(testCase.path)?.description ?? null)
-      : null;
-
-    const results: ExpressionResult[] = [];
-    const errors: string[] = [];
-    for (const expr of libraryResult.expressions) {
-      if (expr.name === 'Error') {
-        errors.push(expr.value);
-      } else {
-        results.push(expr);
-      }
-    }
-
-    const resolvedParams: ParameterEntry[] = parametersConfig
-      ? resolveParameters(parametersConfig, libraryName, libraryVersion, testCaseName ?? undefined)
-      : [];
-
-    const parameters: ResultParameterEntry[] = [
-      ...resolvedParams.map(p => ({ name: p.name, type: p.type, value: p.value, source: p.source! })),
-      ...(libraryResult.usedDefaultParameters ?? []).map(p => ({ name: p.name, value: p.value, source: 'default' as const })),
-    ];
-
-    const result: TestCaseResult = {
-      executedAt: executedAtStr,
-      libraryName,
-      testCaseName,
-      testCaseDescription,
-      parameters,
-      results,
-      errors,
-    };
-
-    const patientId = testCaseName ?? 'no-context';
-    const outputPath = Utils.resolvePath(
-      resultDirectoryPath,
-      libraryName,
-      `TestCaseResult-${patientId}.json`,
-    );
-    fse.outputFileSync(outputPath.fsPath, JSON.stringify(result, null, 2));
-  }
-}
-
-export function formatResponse(response: ExecuteCqlResponse): string {
-  const lines: string[] = [];
-  for (let i = 0; i < response.results.length; i++) {
-    if (i > 0) {
-      lines.push('');
-    }
-    for (const expr of response.results[i].expressions) {
-      lines.push(`${expr.name}=${expr.value}`);
-    }
-  }
-  if (response.logs.length > 0) {
-    lines.push('');
-    lines.push('Evaluation logs:');
-    lines.push(...response.logs);
-  }
-  return lines.join('\n');
-}
-
-export function resolveTestConfigPath(testDirectoryPath: Uri): Uri {
-  const jsoncPath = Utils.resolvePath(testDirectoryPath, 'config.jsonc');
-  if (fs.existsSync(jsoncPath.fsPath)) {
-    return jsoncPath;
-  }
-  return Utils.resolvePath(testDirectoryPath, 'config.json');
-}
-
-export function getCqlPaths(cqlFileUri: Uri): CqlPaths | undefined {
-  const project = CqlSolution.getCurrent().findProjectForUri(cqlFileUri);
-  if (!project) {
-    window.showErrorMessage('Unable to determine path to project root.');
-    return;
-  }
-  const projectDirectoryPath = Uri.file(project.igRoot);
-  const libraryDirectoryPath = Utils.resolvePath(projectDirectoryPath, 'input', 'cql');
-  const testDirectoryPath = Utils.resolvePath(projectDirectoryPath, 'input', 'tests');
-  return {
-    projectDirectoryPath: projectDirectoryPath,
-    libraryDirectoryPath: libraryDirectoryPath,
-    optionsPath: Utils.resolvePath(libraryDirectoryPath, 'cql-options.json'),
-    resultDirectoryPath: Utils.resolvePath(testDirectoryPath, 'results'),
-    terminologyDirectoryPath: Utils.resolvePath(
-      projectDirectoryPath,
-      'input',
-      'vocabulary',
-      'valueset',
-    ),
-    testConfigPath: resolveTestConfigPath(testDirectoryPath),
-    testDirectoryPath: testDirectoryPath,
-  };
-}
-
-export function getExcludedTestCases(
-  libraryName: string,
-  testCasesToExclude: TestCaseExclusion[],
-): Map<string, string> {
-  let excludedTestCases = new Map<string, string>();
-  for (let excludedTestCase of testCasesToExclude) {
-    if (excludedTestCase.library == libraryName) {
-      excludedTestCases.set(excludedTestCase.testCase, excludedTestCase.reason);
-    }
-  }
-  return excludedTestCases;
-}
-
-export function getFhirVersion(cqlContent: string): string | null {
-  // Direct FHIR model declaration: using FHIR version 'x.y.z'
-  const fhirMatch = cqlContent.match(/using\s+(?:FHIR|"FHIR")\s+version\s+'(\d[^']*)'/);
-  if (fhirMatch) {
-    const v = fhirMatch[1];
-    if (v.startsWith('2')) return 'DSTU2';
-    if (v.startsWith('3')) return 'DSTU3';
-    if (v.startsWith('4')) return 'R4';
-    if (v.startsWith('5')) return 'R5';
-  }
-
-  // QICore model declaration: using QICore version 'x.y.z'
-  // QICore 3.x targets FHIR DSTU3; 4.x and above target FHIR R4.
-  const qicoreMatch = cqlContent.match(/using\s+(?:QICore|"QICore")\s+version\s+'(\d[^']*)'/);
-  if (qicoreMatch) {
-    return qicoreMatch[1].startsWith('3') ? 'DSTU3' : 'R4';
-  }
-
-  // USCore model declaration: all versions target FHIR R4.
-  if (/using\s+(?:USCore|"USCore")\s+version\s+'/.test(cqlContent)) {
-    return 'R4';
-  }
-
-  return null;
-}
-
-export function getLibraries(libraryPath: Uri): Array<Uri> {
-  if (!fs.existsSync(libraryPath.fsPath)) {
-    log.warn(`unable to find libraries @ ${libraryPath.fsPath}`);
-    return [];
-  }
-  return glob
-    .sync(`${toGlobPath(libraryPath.fsPath)}/**/*.cql`)
-    .filter(f => fs.statSync(f).isFile())
-    .map(f => Uri.file(f));
-}
-
-
-async function insertLineAtEnd(textEditor: TextEditor, text: string) {
-  const document = textEditor.document;
-  await textEditor.edit(
-    editBuilder => {
-      editBuilder.insert(new Position(document.lineCount, 0), text + '\n');
-    },
-    { undoStopBefore: false, undoStopAfter: false },
-  );
-}
-
-async function waitForTestCasesLoaded(
-  library: CqlLibrary,
-  project: CqlProject,
-  showProgress: boolean,
-): Promise<void> {
-  if (library.testCaseLoadState === 'loaded') return;
-
-  const waitPromise = new Promise<void>(resolve => {
-    const handler = (loaded: CqlLibrary) => {
-      if (loaded === library) {
-        project.off(CqlProjectEvents.LIBRARY_TESTCASES_LOADED, handler);
-        resolve();
-      }
-    };
-    project.on(CqlProjectEvents.LIBRARY_TESTCASES_LOADED, handler);
-    if (library.testCaseLoadState === 'not-loaded') {
-      project.loadTestCasesForLibrary(library).catch(() => {});
-    }
-  });
-
-  if (showProgress) {
-    await window.withProgress(
-      {
-        location: ProgressLocation.Notification,
-        title: `Loading ${library.name}\u2026`,
-        cancellable: false,
-      },
-      () => waitPromise,
-    );
-  } else {
-    await waitPromise;
-  }
-}
-
-export function loadTestConfig(testConfigPath: Uri): TestConfig {
-  if (!fs.existsSync(testConfigPath.fsPath)) {
-    log.info('No test config file found, using default settings', testConfigPath.fsPath);
-    return DEFAULT_TEST_CONFIG;
-  }
-  try {
-    const jsonString = fs.readFileSync(testConfigPath.fsPath, 'utf-8');
-    const errors: ParseError[] = [];
-    const parsed = parseJsonc(jsonString, errors) as TestConfig;
-    if (errors.length > 0) {
-      log.error('Error parsing config file', errors);
-      return DEFAULT_TEST_CONFIG;
-    }
-    return {
-      ...parsed,
-      testCasesToExclude: parsed.testCasesToExclude ?? DEFAULT_TEST_CONFIG.testCasesToExclude,
-      resultFormat: parsed.resultFormat ?? DEFAULT_TEST_CONFIG.resultFormat
-    };
-  } catch (error) {
-    log.error('Error reading config file', error);
-    return { testCasesToExclude: [] };
-  }
-}
-
